Add complete service, mapper, repository, and DTO layers for e-commerce features

+++ src/main/java/co/edu/usbcali/ecommerceusb/service/UserService.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 public interface UserService {
     List<UserResponse> getUsers();
-    UserResponse getUserById (Integer id) throws Exception;
-    UserResponse getUserByEmail (String email) throws Exception;
+    UserResponse getUserById(Long id) throws Exception;
+    UserResponse getUserByEmail(String email) throws Exception;
 
-    UserResponse createUser (CreateUserRequest createUserRequest) throws Exception;
+    UserResponse createUser(CreateUserRequest createUserRequest) throws Exception;
 }
